refactor: move reporter to utilities and rename as logger
